[autothrottle] legacyGetTopicsWithThrottledBrokers
