+++ cmd/autothrottle/topics.go
@@ -3,7 +3,6 @@ package main
 import (
 	"errors"
 	"fmt"
-	"strconv"
 
 	"github.com/DataDog/kafka-kit/v3/kafkazk"
 )
@@ -102,64 +101,15 @@ func (t TopicStates) Filter(fn TopicStatesFilterFn) TopicStates {
 // any topics that have partitions assigned to brokers with a static throttle
 // rate set.
 func (tm *ThrottleManager) getTopicsWithThrottledBrokers() (topicThrottledReplicas, error) {
-	// Fetch all topic states.
-	states, err := tm.getAllTopicStates()
-	if err != nil {
-		return nil, err
+	if !tm.kafkaNativeMode {
+		// Use the direct ZooKeeper config update method.
+		return tm.legacyGetTopicsWithThrottledBrokers()
 	}
-
-	// Lookup brokers with overrides set that are not a reassignment participant.
-	throttledBrokers := tm.brokerOverrides.Filter(notReassignmentParticipant)
 
 	// Construct a topicThrottledReplicas that includes any topics with replicas
 	// assigned to brokers with overrides. The throttled list only includes brokers
 	// with throttles set rather than all configured replicas.
 	var throttleLists = make(topicThrottledReplicas)
 
-	// For each topic...
-	for topicName, state := range states {
-		// TODO(jamie): make this configurable.
-		if topicName == "__consumer_offsets" {
-			continue
-		}
-		// For each partition...
-		for partn, replicas := range state.Partitions {
-			// For each replica assignment...
-			for _, assignedID := range replicas {
-				// If the replica is a throttled broker, append that broker to the
-				// throttled list for this {topic, partition}.
-				if _, exists := throttledBrokers[assignedID]; exists {
-					throttleLists.addReplica(
-						topic(topicName),
-						partn,
-						replicaType("followers"),
-						strconv.Itoa(assignedID))
-				}
-			}
-		}
-	}
-
 	return throttleLists, nil
-}
-
-// getAllTopicStates returns a TopicStates for all topics in Kafka.
-func (tm *ThrottleManager) getAllTopicStates() (TopicStates, error) {
-	var states = make(TopicStates)
-
-	// Get all topics.
-	topics, err := tm.zk.GetTopics(topicsRegex)
-	if err != nil {
-		return nil, err
-	}
-
-	// Fetch state for each topic.
-	for _, topic := range topics {
-		state, err := tm.zk.GetTopicState(topic)
-		if err != nil {
-			return nil, err
-		}
-		states[topic] = *state
-	}
-
-	return states, nil
 }
